Switch to faster sort function.
When placing things in toruses the computation of the cost of a particular
placement solution requires small lists of small integers to be sorted
repeatedly. Previously this was done using qsort but this implementation has
now been replaced with one which uses 'usort'. Usort uses insertion sort for
very small lists and radix sort for longer ones and in both cases doesn't rely
on calling a comparator function. In practice this change results in a speedup
of about 1.7x speedup in placements in torus topologies.