Use shortest-path-star-hop-count not HPWL as cost function.
As a trial, uses the sum of shortest-path hop counts from a net's source to its
sinks as a cost function. This is expected to be advantageous over HPWL
which under-estimates high-fan-out nets and also doesn't account for the
hexagonal nature of SpiNNaker's topology. This new cost function will *over*
estimate net cost in general but should cope much better with higher-fan-out
nets and also hopefully make better use of the network...

Pending some experiments to affirm the quality of placements produced by this
method I plan to make the cost function be computed lazily.