Cache net bounding boxes to improve performance for high-fan-out nets.
Rather than recomputing the bounding box of every impacted net *twice* on every
swap in order to compute the cost of the net, store the bounding box and only
recompute it when it changes. For simplicity this implementation is rather
conservative and actually regenerates the bounding box more often than strictly
necessary.

For low-fan-out nets this has little performance impact. For nets with really
silly fan-outs, I've seen algorithm execution times 2-4x quicker. In a
benchmark for my thesis where I ramp up the fan-out of an input network I find
that this implementation causes runtime to increases sub-linearly with respect
to fan out. The old implementation's runtime was worse-than-linear.